complete authentication

+++ frontend/src/components/WorkoutDetails.js
@@ -1,34 +1,51 @@
-import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
-
+import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 // date fns
-import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutDetails = ({ workout }) => {
-  const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext();
+  const { dispatch } = useWorkoutsContext();
 
   const handleClick = async () => {
+    if(!user){
+      return;
+    }
     const response = await fetch(
       "https://workout-app-qvmk.onrender.com/api/workouts/" + workout._id,
       {
         method: "DELETE",
+        headers: {
+          Authorization: `Bearer ${user.token}`,
+        },
       }
     );
-    const json = await response.json()
+    const json = await response.json();
 
     if (response.ok) {
-      dispatch({type: 'DELETE_WORKOUT', payload: json})
+      dispatch({ type: "DELETE_WORKOUT", payload: json });
     }
-  }
+  };
 
   return (
-    <div className="workout-details">
+    <div className='workout-details'>
       <h4>{workout.title}</h4>
-      <p><strong>Load (kg): </strong>{workout.load}</p>
-      <p><strong>Number of reps: </strong>{workout.reps}</p>
-      <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
-      <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
+      <p>
+        <strong>Load (kg): </strong>
+        {workout.load}
+      </p>
+      <p>
+        <strong>Number of reps: </strong>
+        {workout.reps}
+      </p>
+      <p>
+        {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
+      </p>
+      <span className='material-symbols-outlined' onClick={handleClick}>
+        delete
+      </span>
     </div>
-  )
-}
+  );
+};
 
-export default WorkoutDetails
+export default WorkoutDetails;
